src/pyproject.toml
[project]
name = "fantasy-pitchers"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.30.6",
    "sqlalchemy==2.0.35",
    "alembic==1.13.3",
    "pydantic-settings==2.5.2",
    "mlb-statsapi==1.7.2",
    "apscheduler==3.10.4",
    "python-dotenv==1.0.1",
]
